requirements.txt
streamlit==0.63.1
pandas==1.0.5
numpy==1.19.1
matplotlib==3.3.1
pillow==7.2.0
opencv-contrib-python-headless